+++ cpp/tests/db/find_matches_test.cu
@@ -18,7 +18,9 @@
 #include "db/db_operators.cuh"
 #include "gtest/gtest.h"
 #include "high_res_clock.h"
+#include "rmm/device_buffer.hpp"
 #include "test_utils.h"
+#include "utilities/error_utils.h"
 #include "utilities/graph_utils.cuh"
 
 class Test_FindMatches : public ::testing::Test {
@@ -75,17 +77,17 @@ TEST_F(Test_FindMatches, firstTest)
   p.addEntry(p1);
   p.addEntry(p2);
   p.addEntry(p3);
-  cugraph::db::db_result<int32_t> result = cugraph::db::findMatches(p, table, nullptr, 1);
+  cugraph::db::db_result<int32_t> result =
+    cugraph::db::findMatches<int32_t>(p, table, nullptr, 0, 1);
   ASSERT_EQ(result.getSize(), 1);
-  int32_t* resultA = new int32_t[result.getSize()];
-  int32_t* resultB = new int32_t[result.getSize()];
-  cudaMemcpy(resultA, result.getData("a"), sizeof(int32_t) * result.getSize(), cudaMemcpyDefault);
-  cudaMemcpy(resultB, result.getData("b"), sizeof(int32_t) * result.getSize(), cudaMemcpyDefault);
+  std::vector<int32_t> resultA(result.getSize());
+  std::vector<int32_t> resultB(result.getSize());
+  CUDA_TRY(cudaMemcpy(
+    resultA.data(), result.getData("a"), sizeof(int32_t) * result.getSize(), cudaMemcpyDefault));
+  CUDA_TRY(cudaMemcpy(
+    resultB.data(), result.getData("b"), sizeof(int32_t) * result.getSize(), cudaMemcpyDefault));
   ASSERT_EQ(resultA[0], 1);
   ASSERT_EQ(resultB[0], 2);
-
-  delete[] resultA;
-  delete[] resultB;
 }
 
 TEST_F(Test_FindMatches, secondTest)
@@ -106,23 +108,23 @@ TEST_F(Test_FindMatches, secondTest)
   q.addEntry(q2);
   q.addEntry(q3);
 
-  cugraph::db::db_result<int32_t> result = cugraph::db::findMatches(q, table, nullptr, 2);
+  cugraph::db::db_result<int32_t> result =
+    cugraph::db::findMatches<int32_t>(q, table, nullptr, 0, 2);
 
   std::cout << result.toString();
 
   ASSERT_EQ(result.getSize(), 2);
-  int32_t* resultA = new int32_t[result.getSize()];
-  int32_t* resultB = new int32_t[result.getSize()];
-  cudaMemcpy(resultA, result.getData("a"), sizeof(int32_t) * result.getSize(), cudaMemcpyDefault);
-  cudaMemcpy(resultB, result.getData("b"), sizeof(int32_t) * result.getSize(), cudaMemcpyDefault);
+  std::vector<int32_t> resultA(result.getSize());
+  std::vector<int32_t> resultB(result.getSize());
+  CUDA_TRY(cudaMemcpy(
+    resultA.data(), result.getData("a"), sizeof(int32_t) * result.getSize(), cudaMemcpyDefault));
+  CUDA_TRY(cudaMemcpy(
+    resultB.data(), result.getData("b"), sizeof(int32_t) * result.getSize(), cudaMemcpyDefault));
 
   ASSERT_EQ(resultA[0], 1);
   ASSERT_EQ(resultB[0], 1);
   ASSERT_EQ(resultA[1], 1);
   ASSERT_EQ(resultB[1], 2);
-
-  delete[] resultA;
-  delete[] resultB;
 }
 
 TEST_F(Test_FindMatches, thirdTest)
@@ -139,23 +141,21 @@ TEST_F(Test_FindMatches, thirdTest)
   q.addEntry(q2);
   q.addEntry(q3);
 
-  int32_t* frontier_ptr;
-  cudaMalloc(&frontier_ptr, sizeof(int32_t));
+  rmm::device_buffer frontier(sizeof(int32_t));
+  int32_t* frontier_ptr = reinterpret_cast<int32_t*>(frontier.data());
   thrust::fill(thrust::device, frontier_ptr, frontier_ptr + 1, 0);
-  gdf_column* frontier = (gdf_column*)malloc(sizeof(gdf_column));
-  cugraph::detail::gdf_col_set_defaults(frontier);
-  gdf_column_view(frontier, frontier_ptr, nullptr, 1, GDF_INT32);
 
-  cugraph::db::db_result<int32_t> result = cugraph::db::findMatches(q, table, frontier, 0);
+  cugraph::db::db_result<int32_t> result =
+    cugraph::db::findMatches<int32_t>(q, table, frontier_ptr, 1, 0);
 
   ASSERT_EQ(result.getSize(), 1);
-  int32_t* resultA = new int32_t[result.getSize()];
-  cudaMemcpy(resultA, result.getData("a"), sizeof(int32_t) * result.getSize(), cudaMemcpyDefault);
+  std::vector<int32_t> resultA(result.getSize());
+  CUDA_TRY(cudaMemcpy(
+    resultA.data(), result.getData("a"), sizeof(int32_t) * result.getSize(), cudaMemcpyDefault));
 
   std::cout << result.toString();
 
   ASSERT_EQ(resultA[0], 0);
-  delete[] resultA;
 }
 
 TEST_F(Test_FindMatches, fourthTest)
@@ -174,22 +174,25 @@ TEST_F(Test_FindMatches, fourthTest)
   q.addEntry(q3);
   q.addEntry(q4);
 
-  cugraph::db::db_result<int32_t> result = cugraph::db::findMatches(q, table, nullptr, 0);
+  cugraph::db::db_result<int32_t> result =
+    cugraph::db::findMatches<int32_t>(q, table, nullptr, 0, 0);
   std::cout << result.toString();
   ASSERT_EQ(result.getSize(), 3);
 
-  int32_t* resultA = new int32_t[result.getSize()];
-  cudaMemcpy(resultA, result.getData("a"), sizeof(int32_t) * result.getSize(), cudaMemcpyDefault);
-  int32_t* resultR = new int32_t[result.getSize()];
-  cudaMemcpy(resultR, result.getData("r"), sizeof(int32_t) * result.getSize(), cudaMemcpyDefault);
+  std::vector<int32_t> resultA(result.getSize());
+  std::vector<int32_t> resultR(result.getSize());
+
+  CUDA_TRY(cudaMemcpy(
+    resultA.data(), result.getData("a"), sizeof(int32_t) * result.getSize(), cudaMemcpyDefault));
+  CUDA_TRY(cudaMemcpy(
+    resultR.data(), result.getData("r"), sizeof(int32_t) * result.getSize(), cudaMemcpyDefault));
+
   ASSERT_EQ(resultA[0], 0);
   ASSERT_EQ(resultA[1], 1);
   ASSERT_EQ(resultA[2], 2);
   ASSERT_EQ(resultR[0], 0);
   ASSERT_EQ(resultR[1], 1);
   ASSERT_EQ(resultR[2], 2);
-  delete[] resultA;
-  delete[] resultR;
 }
 
 TEST_F(Test_FindMatches, fifthTest)
@@ -207,22 +210,23 @@ TEST_F(Test_FindMatches, fifthTest)
   q.addEntry(q2);
   q.addEntry(q3);
 
-  cugraph::db::db_result<int32_t> result = cugraph::db::findMatches(q, table, nullptr, 1);
+  cugraph::db::db_result<int32_t> result =
+    cugraph::db::findMatches<int32_t>(q, table, nullptr, 0, 1);
   std::cout << result.toString();
 
   ASSERT_EQ(result.getSize(), 2);
-  int32_t* resultA = new int32_t[result.getSize()];
-  int32_t* resultB = new int32_t[result.getSize()];
-  cudaMemcpy(resultA, result.getData("a"), sizeof(int32_t) * result.getSize(), cudaMemcpyDefault);
-  cudaMemcpy(resultB, result.getData("b"), sizeof(int32_t) * result.getSize(), cudaMemcpyDefault);
+  std::vector<int32_t> resultA(result.getSize());
+  std::vector<int32_t> resultB(result.getSize());
+
+  CUDA_TRY(cudaMemcpy(
+    resultA.data(), result.getData("a"), sizeof(int32_t) * result.getSize(), cudaMemcpyDefault));
+  CUDA_TRY(cudaMemcpy(
+    resultB.data(), result.getData("b"), sizeof(int32_t) * result.getSize(), cudaMemcpyDefault));
 
   ASSERT_EQ(resultA[0], 0);
   ASSERT_EQ(resultA[1], 0);
   ASSERT_EQ(resultB[0], 2);
   ASSERT_EQ(resultB[1], 3);
-
-  delete[] resultA;
-  delete[] resultB;
 }
 
 int main(int argc, char** argv)